requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
et_xmlfile==2.0.0
frozenlist==1.6.0
greenlet==3.2.2
idna==3.10
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.4.4
openpyxl==3.1.5
propcache==0.3.1
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
starlette-admin==0.14.1
typing-inspection==0.4.0
typing_extensions==4.13.2
yarl==1.20.0
